Deleted typeresult : flie restruct
